Move bloodhound package info function to the helpers
