+++ nxc/protocols/ldap.py
@@ -43,6 +43,7 @@ from nxc.protocols.ldap.gmsa import MSDS_MANAGEDPASSWORD_BLOB
 from nxc.protocols.ldap.kerberos import KerberosAttacks
 from nxc.parsers.ldap_results import parse_result_attributes
 from nxc.helpers.ntlm_parser import parse_challenge
+from nxc.helpers.misc import get_bloodhound_info
 
 ldap_error_status = {
     "1": "STATUS_NOT_SUPPORTED",
@@ -1219,64 +1220,6 @@ class ldap(connection):
             self.logger.fail("No string provided :'(")
 
     def bloodhound(self):
-
-        def get_bloodhound_info():
-            """
-            Detect which BloodHound package is installed (regular or CE) and its version.
-            
-            Returns
-            -------
-                tuple: (package_name, version, is_ce)
-                    - package_name: Name of the installed package ('bloodhound', 'bloodhound-ce', or None)
-                    - version: Version string of the installed package (or None if not installed)
-                    - is_ce: Boolean indicating if it's the Community Edition
-            """
-            import importlib.metadata
-            import importlib.util
-            
-            # First check if any BloodHound package is available to import
-            if importlib.util.find_spec("bloodhound") is None:
-                return None, None, False
-            
-            # Try to get version info from both possible packages
-            version = None
-            package_name = None
-            is_ce = False
-            
-            # Check for bloodhound-ce first
-            try:
-                version = importlib.metadata.version("bloodhound-ce")
-                package_name = "bloodhound-ce"
-                is_ce = True
-            except importlib.metadata.PackageNotFoundError:
-                # Check for regular bloodhound
-                try:
-                    version = importlib.metadata.version("bloodhound")
-                    package_name = "bloodhound"
-                    
-                    # Even when installed as 'bloodhound', check if it's actually the CE version
-                    if version and ("ce" in version.lower() or "community" in version.lower()):
-                        is_ce = True
-                except importlib.metadata.PackageNotFoundError:
-                    # No bloodhound package found via metadata
-                    pass
-
-            # In case we can import it but metadata is not working, check the module itself
-            if not version:
-                try:
-                    import bloodhound
-                    version = getattr(bloodhound, "__version__", "unknown")
-                    package_name = "bloodhound"
-                    
-                    # Check if it's CE based on version string
-                    if "ce" in version.lower() or "community" in version.lower():
-                        is_ce = True
-                        package_name = "bloodhound-ce"
-                except ImportError:
-                    pass
-
-            return package_name, version, is_ce
-
         # Check which version is desired
         use_bhce = self.config.getboolean("BloodHound-CE", "bhce_enabled", fallback=False)
         package_name, version, is_ce = get_bloodhound_info()
